simplifying names
